react-form/server/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
mysql-connector-python==8.0.26 